refactor and add test
