+++ examples/pinocchio/src/lib.rs
@@ -1,3 +1,9 @@
 pub mod common;
 pub mod qap;
 pub mod r1cs;
+pub mod prover;
+pub mod setup;
+pub mod verifier;
+
+// Re-export the test_utils module for use in integration test
+pub mod test_utils;
